Fixes #56
