+++ lib/functions.js
@@ -7,14 +7,6 @@
 
 const fs = require('fs');
 
-const {
-	milliseconds, seconds, minutes,
-	hours, days, weeks,
-	months, years, decades, centuries,
-} = require("../data/time.js");
-
-const timeMap = require("../data/time.js");
-
 
 /**
  * get a random number between specified values
@@ -122,167 +114,6 @@ function writeFile(path, content)
 	}
 }
 
-
-/**
- * break a time stamp (in milliseconds) down to a count of each components, eg. days/hours/mins/secs/ms
- *
- * @param  Number
- * @param  Object
- * @return Boolean
- */
-function formatTimeStamp(ts, options)
-{
-	let components = [],
-		componentCount = 0,
-		description = "",
-		sep = "",
-
-		tsTemp = ts,
-
-		largest = null;
-
-		o = {
-			showCenturies: true,
-			showDecades: true,
-			showYears: true,
-			showMonths: true,
-			showWeeks: true,
-			showDays: true,
-			showHours: true,
-			showMinutes: true,
-			showSeconds: true,
-			showMilliseconds: false,
-		},
-
-		componentLabels = {
-			centuries: "century",
-			decades: "decade",
-			years: "year",
-			months: "month",
-			weeks: "week",
-			days: "day",
-			hours: "hour",
-			minutes: "minute",
-			seconds: "second",
-			milliseconds: "millisecond",
-		},
-
-		data = {
-			centuries: 0,
-			decades: 0,
-			years: 0,
-			months: 0,
-			weeks: 0,
-			days: 0,
-			hours: 0,
-			minutes: 0,
-			seconds: 0,
-			milliseconds: 0,
-		},
-
-		totals = copyObject(data);
-
-	if (typeof ts !== "number" ||
-		isNaN(ts) === true ||
-		ts < milliseconds) {
-		return;
-	}
-
-	// merge options
-	o = { ...o, ...options };
-
-	// dissect the time stamp
-	for (const [k, v] of Object.entries(totals)) {
-		const sv = constructShowVar(k);
-
-		if (ts < timeMap[k]) {
-			continue;
-		}
-
-		if (largest === null) {
-			largest = k;
-		}
-
-		totals[k] = Math.floor(ts/timeMap[k]);
-
-		// eg. showCenturies
-		if (o[sv] === true) {
-			data[k] = Math.floor(tsTemp/timeMap[k]);
-			tsTemp -= data[k]*timeMap[k];
-
-			// check for plural
-			if (data[k] > 1) {
-				componentLabels[k] = k;
-			}
-
-			components.push(`${data[k]} ${componentLabels[k]}`);
-		}
-
-		// done? then be done
-		if (tsTemp <= 0) {
-			break;
-		}
-	}
-
-	if (components.length === 0) {
-		console.log("no output: formatTimeStamp", ts, data, totals);
-
-		return false;
-	}
-
-	// build the description
-	for (const c of components) {
-		description += `${sep}${c}`;
-
-		sep = getSeparator(componentCount, components.length);
-		componentCount++;
-	}
-
-	return {
-		description: description,
-		data: data,
-		totals: totals,
-		largest: largest,
-	};
-}
-
-/**
- * build the element display option var eg. o.showCenturies
- *
- * @param  Object
- * @return String
- */
-function constructShowVar(c)
-{
-	const l = c.slice(0, 1).toUpperCase();
-	const remainder = c.slice(1);
-
-	return `show${l}${remainder}`;
-}
-
-/**
- * get the correct separator for a string item list
- *
- * @param  Number
- * @param  Number
- * @return String
- */
-function getSeparator(current, total)
-{
-	let sep = ", ";
-
-	if (total === 2) {
-		sep = " and ";
-	} else {
-		if (total > 2 &&
-			(total-1)-current === 1) {
-			sep = ", and ";
-		}
-	}
-
-	return sep;
-}
-
 /**
  * check subscription status using tmi.js userstate info
  *
@@ -304,5 +135,5 @@ function checkSub(userstate)
 
 
 module.exports = {
-	rnd, arnd, readFile, writeFile, formatTimeStamp, copyObject, getSeparator, checkSub,
+	rnd, arnd, readFile, writeFile, copyObject, checkSub,
 };
