+++ lib/twitch-chat-bot.js
@@ -33,12 +33,14 @@ const {
 	months, years, decades, centuries,
 } = require("../data/time.js");
 
+const timeMap = require("../data/time.js");
+
 /* internal libraries */
 
 const TwitchChatBotErrorHandler = require("./error-handler.js");
 
 const {
-	checkSub,
+	copyObject, checkSub,
 } = require("./functions.js");
 
 /* external libraries */
@@ -1189,6 +1191,166 @@ class TwitchChatBot {
 	sendMessage(text, options)
 	{
 		this.messaging.queueMessage(text, options);
+	}
+
+	/**
+	 * break a time stamp (in milliseconds) down to a count of each components, eg. days/hours/mins/secs/ms
+	 *
+	 * @param  Number
+	 * @param  Object
+	 * @return Boolean
+	 */
+	formatTimeStamp(ts, options)
+	{
+		let components = [],
+			componentCount = 0,
+			description = "",
+			sep = "",
+
+			tsTemp = ts,
+
+			largest = null,
+
+			o = {
+				showCenturies: true,
+				showDecades: true,
+				showYears: true,
+				showMonths: true,
+				showWeeks: true,
+				showDays: true,
+				showHours: true,
+				showMinutes: true,
+				showSeconds: true,
+				showMilliseconds: false,
+			},
+
+			componentLabels = {
+				centuries: "century",
+				decades: "decade",
+				years: "year",
+				months: "month",
+				weeks: "week",
+				days: "day",
+				hours: "hour",
+				minutes: "minute",
+				seconds: "second",
+				milliseconds: "millisecond",
+			},
+
+			data = {
+				centuries: 0,
+				decades: 0,
+				years: 0,
+				months: 0,
+				weeks: 0,
+				days: 0,
+				hours: 0,
+				minutes: 0,
+				seconds: 0,
+				milliseconds: 0,
+			},
+
+			totals = copyObject(data);
+
+		if (typeof ts !== "number" ||
+			isNaN(ts) === true ||
+			ts < milliseconds) {
+			return;
+		}
+
+		// merge options
+		o = { ...o, ...options };
+
+		// dissect the time stamp
+		for (const [k, v] of Object.entries(totals)) {
+			const sv = this.constructShowVar(k);
+
+			if (ts < timeMap[k]) {
+				continue;
+			}
+
+			if (largest === null) {
+				largest = k;
+			}
+
+			totals[k] = Math.floor(ts/timeMap[k]);
+
+			// eg. showCenturies
+			if (o[sv] === true) {
+				data[k] = Math.floor(tsTemp/timeMap[k]);
+				tsTemp -= data[k]*timeMap[k];
+
+				// check for plural
+				if (data[k] > 1) {
+					componentLabels[k] = k;
+				}
+
+				components.push(`${data[k]} ${componentLabels[k]}`);
+			}
+
+			// done? then be done
+			if (tsTemp <= 0) {
+				break;
+			}
+		}
+
+		if (components.length === 0) {
+			console.log("no output: formatTimeStamp", ts, data, totals);
+
+			return false;
+		}
+
+		// build the description
+		for (const c of components) {
+			description += `${sep}${c}`;
+
+			sep = this.getSeparator(componentCount, components.length);
+			componentCount++;
+		}
+
+		return {
+			description: description,
+			data: data,
+			totals: totals,
+			largest: largest,
+		};
+	}
+
+	/**
+	 * build the element display option var eg. o.showCenturies
+	 *
+	 * @param  Object
+	 * @return String
+	 */
+	constructShowVar(c)
+	{
+		const l = c.slice(0, 1).toUpperCase();
+		const remainder = c.slice(1);
+
+		return `show${l}${remainder}`;
+	}
+
+	/**
+	 * get the correct separator for a string item list
+	 *
+	 * @param  Number
+	 * @param  Number
+	 * @return String
+	 */
+	getSeparator(current, total)
+	{
+		let sep = ", ";
+
+		if (total === 2) {
+			sep = " and ";
+		} else {
+			if (total > 2 &&
+				(total - 1) - current === 1) {
+				sep = ", and ";
+			}
+		}
+
+		return sep;
 	}
